Enhance CBOW model and tokenizer setup
This commit updates the CBOW model by increasing the embedding and linear layer dimensions to 256, modifies the learning rate for the optimizer, and introduces a new Regressor class for further processing. Additionally, it refines the NLTK setup by creating a custom directory for NLTK data downloads and adjusts the tokenizer scripts to utilize this new setup, ensuring a more organized and efficient tokenization process.